final big update
