+++ BusinessCard.pass/pass.json
@@ -8,12 +8,12 @@
 
     "foregroundColor": "rgb(255, 255, 255)",
     "backgroundColor": "rgb(0, 0, 0)",
-    "organizationName": "Test company",
+    "organizationName": "TestTestCompany LLC",
     "description": "business card",
     "logoText": "",
     "barcode": {
         "altText": "add my contact (My MeCard)",
-        "message": "MECARD:N:Lname,Fname;EMAIL:a@b.c;TEL:9876543210;BDAY:19990526;;",
+        "message": "MECARD:N:Lname,Fname;EMAIL:a@b.c;TEL:9876543210;BDAY:yyyymmdd;;",
         "format": "PKBarcodeFormatQR",
         "messageEncoding": "iso-8859-1"
     },
@@ -73,12 +73,12 @@
             {
                 "key": "date-of-birth-full",
                 "label": "date of birth",
-                "value": "19990526"
+                "value": "yyyymmdd"
             },
             {
                 "key": "time-zone",
                 "label": "time zone",
-                "value": "+0100"
+                "value": "UTC -5"
             },
             {
                 "key": "paypal-account",
@@ -91,23 +91,8 @@
                 "value": ""
             },
             {
-                "key": "instagram-account",
-                "label": "Instagram",
-                "value": ""
-            },
-            {
-                "key": "snapchat-account",
-                "label": "Snapchat",
-                "value": ""
-            },
-            {
                 "key": "twitter-account",
                 "label": "Twitter",
-                "value": ""
-            },
-            {
-                "key": "telegram-account",
-                "label": "Telegram",
                 "value": ""
             },
             {
@@ -118,11 +103,6 @@
             {
                 "key": "linkedin-account",
                 "label": "LinkedIn",
-                "value": ""
-            },
-            {
-                "key": "lichess-account",
-                "label": "Lichess",
                 "value": ""
             }
         ]
